requirements.txt
httpx
firebase-admin
pandas
google-cloud-bigquery[pyarrow]
